requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
asyncio==4.0.0
    # via -r requirements.in
attrs==25.3.0
    # via aiohttp
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via yarl
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
psycopg[binary,pool]==3.2.10
    # via -r requirements.in
psycopg-binary==3.2.10
    # via psycopg
psycopg-pool==3.2.6
    # via psycopg
python-dotenv==1.1.1
    # via -r requirements.in
typing-extensions==4.15.0
    # via psycopg-pool
yarl==1.20.1
    # via aiohttp
